Time: 0 ms (100%), Space: 7.9 MB (18.82%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -4,8 +4,8 @@ public:
         int low=1;
         int high = num;
         while(low<=high){
-            long long mid= low + (high-low)/2;
-            long long square = mid*mid;
+            long mid= low + (high-low)/2;
+            long  square = mid*mid;
             if(square==num){
                 return true;
             }
